remove the unnecessary vulkan_utils package (#1149)
- remove the unnecessary vulkan_utils package
- Separate the external API and xml parsing
