+++ vulkan_generator/vulkan_parser/BUILD.bazel
@@ -16,12 +16,28 @@ load("@rules_python//python:defs.bzl", "py_library")
 load("//tools/build:rules.bzl", "py_lint")
 
 py_library(
+    name = "vulkan_parser_internal",
+    srcs = glob(
+        ["*.py"],
+        exclude = [
+            "types.py",
+            "parser.py",
+        ],
+    ),
+    srcs_version = "PY3",
+    visibility = ["//visibility:private"],
+)
+
+py_library(
     name = "vulkan_parser",
-    srcs = glob(["*.py"]),
+    srcs = [
+        "parser.py",
+        "types.py",
+    ],
     srcs_version = "PY3",
     visibility = ["//visibility:public"],
     deps = [
-        "//vulkan_generator/vulkan_utils",
+        "vulkan_parser_internal",
     ],
 )
