--- vulkan_generator/vulkan_utils/BUILD.bazel
@@ -1,31 +1,0 @@
-# Copyright (C) 2022 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-load("@rules_python//python:defs.bzl", "py_library")
-load("//tools/build:rules.bzl", "py_lint")
-
-py_library(
-    name = "vulkan_utils",
-    srcs = glob(["*.py"]),
-    srcs_version = "PY3",
-    visibility = ["//visibility:public"],
-)
-
-py_lint(
-    name = "lint",
-    srcs = glob(["*.py"]),
-    deps = [
-        "vulkan_utils",
-    ],
-)
